requirements.txt
faiss-cpu==1.13.1
langchain==1.1.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.2
langchain-openai==1.1.1
langchain-text-splitters==1.0.0
python-dotenv==1.2.1
streamlit==1.52.1
tiktoken==0.12.0

